+++ packages/flutter_solidart/example/lib/main.dart
@@ -2,12 +2,12 @@ import 'dart:developer' as dev;
 
 import 'package:example/pages/counter.dart';
 import 'package:example/pages/derived_signal.dart';
-import 'package:example/pages/dual_signal_builder.dart';
 import 'package:example/pages/effects.dart';
 import 'package:example/pages/map_signal.dart';
 import 'package:example/pages/resource.dart';
 import 'package:example/pages/set_signal.dart';
 import 'package:example/pages/show.dart';
+import 'package:example/pages/signal_builder.dart';
 import 'package:example/pages/solid/solid.dart';
 import 'package:example/pages/list_signal.dart';
 import 'package:flutter/material.dart';
@@ -63,7 +63,7 @@ final routes = <String, WidgetBuilder>{
   '/show': (_) => const ShowPage(),
   '/derived-signal': (_) => const DerivedSignalsPage(),
   '/effects': (_) => const EffectsPage(),
-  '/dual-signal-builder': (_) => const DualSignalBuilderPage(),
+  '/signal-builder': (_) => const SignalBuilderPage(),
   '/resource': (_) => const ResourcePage(),
   '/solid': (_) => const SolidPage(),
   '/list-signal': (_) => const ListSignalPage(),